correctly handle an empty `mo-git-blame-git-blame-args'
Recently, support was added for specifying custom arguments to be passed
to `git blame' [657dffd: "Support passing extra arguments to git
blame."]. However, it appears that when a user doesn't specify
anything (i.e. leaves the defaults of "") everything is broken.

Fix this by filtering out "" from the arguments being passed to
`mo-git-blame-run*'.